+++ common/hydracore/call_server.go
@@ -7,8 +7,8 @@ const (
 	callClientEnabled = false
 	callServerEnabled = true
 	distributionRole  = "vps"
-	callWireMin       = 5
-	callWireMax       = 5
+	callWireMin       = 6
+	callWireMax       = 6
 )
 
 var (